requirements.txt
fastapi
jinja2
uvicorn
python-multipart
python-jose
passlib
filelock
